refactor: 페이지 파일 분할과 절대경로 적용
